only run C spec compiled with Clang
In order to make sure we have valid failures on this branch, and not
just a bunch of false negatives, we're only running the C specs, and
building with clang. Eventually, we'll merge in support for gcc and more
of the test suite.

For now, we want to only include the specs we things should be passing