add support for Maybe APIs.
This adds some conversion classes to automatically unbox both
`Maybe<T>`, and `MaybeLocal<T>` values. In general, you take the
conversion class, and access the `::Maybe` class underneath it. For
example, to unbox values of `v8::Maybe<v8::String>`, you'd use
`String::Maybe()`

This PR doesn't add *all* the helpers, but some and an example in the
implementation of the `v8::String.ToString()` method.