Change names access helper to iterate over the ancestors instead of doing dup/shift
Due to setup of the helper calling the get or set methods on any object bound to a ruby object will call the accessibe_names each time. This will result in the creation of an new array which will be used for iteration and then discarded, which is quite inefficient if you plan on calling into ruby often.